+++ zsh/.functions.d/03-archives.zsh
@@ -19,7 +19,7 @@ extract() {
             *.tar) tar xf "$1" ;;
             *.tbz2) tar xjf "$1" ;;
             *.tgz) tar xzf "$1" ;;
-            *.zip) unzip "$1" ;;
+            *.zip) unzip "$1" -x "__MACOSX/*" ;;
             *.Z) uncompress "$1" ;;
             *.7z) 7z x "$1" ;;
             *) echo "'$1' cannot be extracted via extract()" ;;